backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
gunicorn==21.2.0
